Backend/requirements.txt
Flask
flask-smorest
python-dotenv
pyrebase4
flask-cors
firebase-admin
requests>=2.26.0
pytest
pytest-flask
openai==0.28